Fix Int64 cast in LeaderboardGenerator and serialization crash in auth handler
- LeaderboardGenerator: SQL ROW_NUMBER() returns bigint but code used
  GetInt32() causing InvalidCastException in production (#2893). Changed
  to GetInt64() with int cast.
- UserIsValidUserAuthHandler: catch block serialized ClaimsPrincipal via
  JsonConvert which triggers PlatformNotSupportedException (#2887).
  Replaced with simple claims string join.

Closes #2893
Closes #2887

+++ TrashMobDailyJobs/LeaderboardGenerator.cs
@@ -87,7 +87,7 @@ namespace TrashMobDailyJobs
                 var region = reader.IsDBNull(2) ? null : reader.GetString(2);
                 var city = reader.IsDBNull(3) ? null : reader.GetString(3);
                 var score = reader.GetDecimal(4);
-                var rank = reader.GetInt32(5);
+                var rank = (int)reader.GetInt64(5);
 
                 // Insert Global entry
                 await InsertLeaderboardEntry(conn, "User", userId, userName, leaderboardType, timeRange, "Global", null, score, rank, computedDate);
@@ -199,7 +199,7 @@ ORDER BY Rank";
                 var region = reader.IsDBNull(2) ? null : reader.GetString(2);
                 var city = reader.IsDBNull(3) ? null : reader.GetString(3);
                 var score = reader.GetDecimal(4);
-                var rank = reader.GetInt32(5);
+                var rank = (int)reader.GetInt64(5);
 
                 // Insert Global entry
                 await InsertLeaderboardEntry(conn, "Team", teamId, teamName, leaderboardType, timeRange, "Global", null, score, rank, computedDate);
